requirements.txt
requests
BeautifulSoup4
pandas
pdfplumber
openpyxl
docx 
python-pptx
urllib3
certifi 
webdriver-manager
PyYAML
lxml